Android 16 KB support
