+++ android/.gitignore
@@ -1,6 +1,7 @@
 # Gradle files
 .gradle/
 build/
+.kotlin
 
 # Local configuration file (sdk path, etc)
 local.properties
