+++ lib/localizations.dart
@@ -1,5 +1,5 @@
+import 'l10n/app_localizations.dart' as intl;
 import 'package:flutter/material.dart';
-import 'package:flutter_gen/gen_l10n/app_localizations.dart' as intl;
 import 'package:slidemix/main.dart';
 
 // flutter gen-l10n
@@ -20,7 +20,8 @@ class AppLocalizations {
   static List<LocalizationsDelegate<dynamic>> get localizationsDelegates =>
       intl.AppLocalizations.localizationsDelegates;
 
-  static List<Locale> get supportedLocales => intl.AppLocalizations.supportedLocales;
+  static List<Locale> get supportedLocales =>
+      intl.AppLocalizations.supportedLocales;
 }
 
 extension AppLocalizationsX on intl.AppLocalizations {
