+++ pubspec.yaml
@@ -17,10 +17,10 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 # In Windows, build-name is used as the major, minor, and patch parts
 # of the product and file versions while build-number is used as the build suffix.
-version: 2.0.4+20004
+version: 2.0.5+20005
 
 environment:
-  sdk: '>=3.4.0 <4.0.0'
+  sdk: '>=3.8.0 <4.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -34,27 +34,23 @@ dependencies:
   flutter_localizations:
     sdk: flutter
 
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: 1.0.8
-
   # https://pub.dev/packages/equatable
   equatable: 2.0.7
 
   # https://pub.dev/packages/firebase_analytics
-  firebase_analytics: 11.4.0
+  firebase_analytics: 11.5.2
 
   # https://pub.dev/packages/firebase_core
-  firebase_core: 3.10.0
+  firebase_core: 3.15.1
 
   # https://pub.dev/packages/firebase_crashlytics
-  firebase_crashlytics: 4.3.0
+  firebase_crashlytics: 4.3.9
 
   # https://pub.dev/packages/flutter_bloc
-  flutter_bloc: 8.1.6
+  flutter_bloc: 9.1.1
 
   # https://pub.dev/packages/flutter_native_splash
-  flutter_native_splash: 2.4.4
+  flutter_native_splash: 2.4.6
 
   # https://pub.dev/packages/floor
   floor: 1.5.0
@@ -66,34 +62,34 @@ dependencies:
   image_picker: 1.1.2
 
   # https://pub.dev/packages/intl
-  intl: 0.19.0
+  intl: 0.20.2
 
   # https://pub.dev/packages/logger
-  logger: 2.5.0
+  logger: 2.6.0
 
   # https://pub.dev/packages/path
-  path: 1.9.0
+  path: 1.9.1
 
   # https://pub.dev/packages/path_provider
   path_provider: 2.1.5
 
   # https://pub.dev/packages/share_plus
-  share_plus: 10.1.3
+  share_plus: 11.0.0
 
   # https://pub.dev/packages/shared_preferences
-  shared_preferences: 2.3.5
+  shared_preferences: 2.5.3
 
   # https://pub.dev/packages/sizer
   sizer: 3.0.5
 
   # https://pub.dev/packages/sqflite
-  sqflite: 2.4.1
+  sqflite: 2.4.2
 
   # https://pub.dev/packages/timeago
-  timeago: 3.7.0
+  timeago: 3.7.1
 
   # https://pub.dev/packages/video_player
-  video_player: 2.9.2
+  video_player: 2.10.0
 
 dev_dependencies:
   # The "flutter_lints" package below contains a set of recommended lints to
@@ -106,7 +102,7 @@ dev_dependencies:
   # https://pub.dev/packages/floor
   floor_generator: 1.5.0
   # flutter packages pub run build_runner build
-  build_runner: 2.4.14
+  build_runner: 2.5.4
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
